Twitoff/requirements.txt
basilica==0.2.5
certifi==2018.8.24
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.8
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
numpy==1.16.2
oauthlib==3.0.1
Pillow==5.4.1
pipenv==2018.11.26
psycopg2==2.7.7
PySocks==1.6.8
python-decouple==3.1
python-dotenv==0.10.1
requests==2.21.0
requests-oauthlib==1.2.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
SQLAlchemy==1.2.11
tweepy==3.7.0
urllib3==1.24.1
virtualenv==16.4.0
virtualenv-clone==0.5.1
Werkzeug==0.14.1
